Move OptionsService into services folder
